Cleanup: move CypressCoordinatesHelper back to common/Jore
