+++ ui/src/components/common/Jore/CypressCoordinatesHelper.tsx
