+++ ui/src/components/common/Jore/index.ts
@@ -1,3 +1,4 @@
+export * from './CypressCoordinatesHelper';
 export * from './ObservationDateControl';
 export * from './OpenDefaultMapButton';
 export * from './PageTitle';
